chore(explore): 게시물 관련 JS 파일명 변경 및 HTML에 반영
