+++ html/explore.html
@@ -34,6 +34,6 @@
     <div id="footer-placeholder"></div>
     <script src="../js/IncludeFooter.js"></script>
 
-    <script type="module" src="../js/explorePost.js"></script>
+    <script type="module" src="../js/explore.js"></script>
   </body>
 </html>
